Refactor nsmGpuChassis to nsmChassis
This commit entails several changes to refactor the nsmGpuChassis
directory into nsmChassis. The rationale for this change is to allow
the classes to be utilized for other types of chassis, such as Baseboard.

Changes include:
- Renaming the nsmGpuChassis directory to nsmChassis.
- Replacing instances of "NSM_GPU_Chassis" with "NSM_Chassis".
- Moving interface.hpp to the nsmd directory for common usage.
- Moving nsmSensorHelper.hpp/cpp functions to the nsmDevice.hpp/cpp for
common usage.
This refactoring enhances code organization and prepares for future
scalability.

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-11819