Fix unsafe call to roundRobinSensors::front()
With the new changes from !203, Static sensors are now queued in with the
round robin sensors in order to fix failed updates. The logic involves
not queuing succesfully updated static sensors. This means that if the
round-robin queue only consists of static sensors after some time when
the queue is empty, and the service ready logic to mark the device ready
and check the other devices leads to a call to std::queue.front() on
an empty round-robin queue. (In short there is a possibility that the
round-robin queue can be empty in the midst of the polling logic). This
commit fixes that behaviour.