Mute fetchSpeedAndCalcBandwidth function for nsmSwitch
Fixes NVBUG https://nvbugs/4748673

Description:
This fetchSpeedAndCalcBandwidth is very heavy on compute side.
There are a lot of dbus calls with O(n^3) time complexity.

There is large number of ports for nvswitch. For each one
we are calling this function in polling loop.

This makes nsmd non responsive for about 25 seconds.
after that loop continues. We cant say it is a hang but
non responsiveness for short interval of time.

During this period even busctl tree command times out on nsmd.
becuase it is busy doing this heavy computation.

EFFECT ON PLDM:
watchdog timeout is around 1 minute in pldm.
Now if repeated update() is called with fetchSpeedAndCalcBandwidth()
28+32+18 = 78 seconds , then watchdog will make pldm restart.

if repeated it is not called timeout will not occur in pldm.
Created bug to track this: https://nvbugswb.nvidia.com/NVBugs5/redir.aspx?url=/4749722

Testing:
Ran for 30 minutes,
No unresponsiveness seen nsmd, no crash with pldmd