Throttle static sensors
Static Sensors are added to the RR queue for updates in order to have a
retry mechanism for failed updates. Once a static sensor is successfully
updated it is popped from the queue.

With the throttling implementation we wanted to remove the static sensor
from the queue as quickly as possible, hence, they were not throttled.

But this is resulting in very frequent failure logs in the journal.

This patch throttles all sensors. Removing the static sensor check.

Fixes nvbug https://

---

Signed-off-by: akurdunkar <akurdunkar@nvidia.com>