Fix sensor polling logic for Non-priority devices
Fixes nvbug https://nvbugspro.nvidia.com/bug/4908995

---

Non-priority devices: Devices with no priority sensors. (NvSwitches,
CX7, ERoTs)

---

With nsmd!347 the RR refresh limit was set to 5 seconds. nsmd!358 set
the polling time for Non-priority devices equivalent to the RR refresh
limit. This ensured that we weren't crowding the event loop with timer
and socket events and helped in reaching the <300ms freshness criteria
for SMBus when the RR refresh limit was 5s. But as the RR refresh limit
was bumped to 30s, NvSwitch's RR sensors reported and abnormal refresh
intervals of >6minutes [0]. This was because we were refreshing the RR
sensors in 150ms and then had to wait for 30seconds for the update. With
~100 sensors the polling interval really added up and lead to such high
SRIs.

As the RR refresh limit is now bumped to 30s we can afford having the
same polling interval for Non-priority devices. This patch makes the
said changes.

Newly observed RR Refresh Intervals (Worst Case) (1hr Run) [1]

```
GPUs: 31.141s
FPGA: 30.742s
NvSwitches: 31.157s
CX7: 30.646s
FPGA ERoT: 31.015s
CX7 ERoT: 30.570s
```

---

[0]: https://nvidia-my.sharepoint.com/:x:/g/personal/akurdunkar_nvidia_com1/EQyFrCKN_uxHlnwwjD5BEaMBZbulkIdVr7rZrmIV7WV3Ow?e=ovCB8O
[1]: https://nvidia-my.sharepoint.com/:x:/g/personal/akurdunkar_nvidia_com1/EZiH9w7sQwxJs39wu7sDERoBQSL4l1yQpRvM2f7fuw3mLA?e=bJgP9l

---

Signed-off-by: akurdunkar <akurdunkar@nvidia.com>