Fix usage of free in NsmBuildType implementation
This commit is to fix wrong usage of free
in NsmBuildType in a case when function
decode_nsm_query_get_erot_state_parameters_resp
returns with cc set to a value, which is not a success.

It also makes sure that the local variable erotInfo
is initialized with all zeroes.

Tested on a real machine, where cc=1 error is returned.

Fixes nvbug https://nvbugspro.nvidia.com/bug/4914193