Type-6: Support for empty key indices for cancellation
GPU key revocation is not immediate and values
will be in pending state. Cancellation of pending
values can be done by sending the current bits in
EFUSE. When none of the keys are revoked empty
indices needs to be passed to cancel the pending
values. For that check in nsmd needs to be relaxed
so that all bits can be unset.

This patch fixes the issue by allowing empty
indices as a valid keys in key revocation method.

Unit Tests:
  1. nsmtool: Verify empty key indices
  2. nsmtool: Verify indices with keys
  3. Redfish: Verify setting with keys
  4. Redfish: Verify with empty key indices

Fixes nvbug https://nvbugspro.nvidia.com/bug/4900508

Signed-off-by: Chinmay Shripad Hegde <chinmays@nvidia.com>