Type-6: Fix race condition in callback handling
Device sensors are added to the loop before all
the slot objects are created, which is leading to
a race condition where only one slot object in
D-Bus and response handler is trying to update the
slot object which is not present leading to a
crash.

This patch moves device sensor register once all
slot objects are created and also adds a size
check for slot object.

Unit Tests:
  1. Verify no crash is seen over multiple
     iteration
  2. Sanity: Image slot verification ERoT & IRoT
  3. Sanity: Irreversible config
  4. Sanity: UpdateMinSecVersion
  5. Sanity: UpdateCodeAuthentication

Fixes nvbug https://nvbugspro.nvidia.com/bug/4914193

Signed-off-by: Chinmay Shripad Hegde <chinmays@nvidia.com>