Deleting PCIeRetimerPortSensors for QueryScalarGroupTelemetry
This implementation deletes PCIeRetimerPortSensors for
QueryScalarGroupTelemetry for Group2 and Group4

As, we know that PCIeRetimerPorts don’t have the hardware support for
these telemetry, we delete them.

Tested : I tested the changes by doing a FW Update including this MR
and the MR for bmcWeb and PDI. Verified that the redfish URI gives
Null for the unsupported properties for PCIeRetimers Ports.

Fixes nvbug https://nvbugspro.nvidia.com/bug/4879269
Signed-off-by: ￼ <kshitizk@nvidia.com>