feat: Add MNNVLinkTopology to GPU
Situation:
- help to show the tray, module, and IB information.
The information can be obtained from CBC FRU in the host BMC,
but the partners may or may not implement the feature.
So, we rely on GPU via NSM to provide this information.

Change:
- the feature is not about GPU processor instead of the rack/tray id,
so we put the change to deviceManager. in addition, it will run
when the device type is GPU.

- chassis_serial_number will return ff from a GPU when data is invalid.
so we catch the exception and then handle it.

- make nsmtool support to get MNNVLinkTopology data.

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-13222