[gb200-nvl] Temporary fix to block invalid NSM messages to FPGA
Added WaR back in to test SatMC message SPAM:
[INFO]: (0x0000000811400101)
0x01 (TX_ARB)
0x01 (INVALID_EID)
0x40 (PKTSRC,BUSNUM)
0x11 (TARG_EID)
0x08 (SRC_EID)

Issue:
The FPGA on gb200-nvl systems is getting several invalid messages.
These messages are not malformed, they are supported on other
platforms, but return failure completion codes. The NSM service
handles this but it fills up the FPGA error FIFO.

Fix:
This is a temporary WaR for our upcoming build, but we will need
to clean up properly later. The fix adds checks and will fail
out if an invalid message is sent to an FPGA EID. Jira for
long term fix:
https://jirasw.nvidia.com/browse/DGXOPENBMC-14750

Tested:
Build NSMD, loaded on system, and checked FPGA error FIFO with
the mctp_err_dump utility

Without WaR we were seeing error messages build up with the
host off
With WaR we are not (we get a few messages on host power
off, but not a constant stream)

Fixes NVBug https://nvbugspro.nvidia.com/bug/4776251

Signed-off-by: Daniel McClung's avatar <dmcclung@nvidia.com>