nsmd: NsmDevice keeps in offline even its EID back to available
If the MCTP control daemon sends an EID Enabled=false signal when the
sensor polling task is not in the sleep coroutine, the sensor polling
task will stop, and the NsmDevice will remain in an offline state even
if NSMD receives an EID Enabled=true signal. This occurs because setting
the NsmDevice to online status is part of the sensor polling loop.
Therefore, to ensure that the NsmDevice becomes online, the sensor
polling task should continue running even when the device is offline.

Fixes nvbug https://nvbugs/5090875

Signed-off-by: Gilbert Chen <gilbertc@nvidia.com>