Bypass command support check for raw commands
Description:
Adds a new optional parameter to the SendRecvNsmMsg method in
SensorManager to bypass the command support check. This allows the
NsmRawCommandHandler to send any command regardless of whether it's
listed in the command matrix.

Raw commands are meant to provide a direct path for sending any
arbitrary command to devices, so they should not be subject to the same
command matrix restrictions as regular commands used by sensors.

Fixes NVBUG https://nvbugswb.nvidia.com/NVBugs5/redir.aspx?url=/5156096
Testing:
- PID loop working correctly.
- even if driver unloaded for gpu, command should go to device
- check command when device is inactive