Implement persistent mode handling in MockupResponder
This commit introduces a new function `pmDisabled()` to check for
the presence of a specific file that indicates whether persistent
mode is disabled. Additionally, it updates various response
handlers in the `MockupResponder` class to return an
unsupported response when persistent mode is disabled, ensuring
that commands related to platform environmental settings are
handled appropriately.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5236922
Signed-off-by: Paweł Iwaneczko <piwaneczko@nvidia.com>