Refactor, coroutine handling and improve README instructions
This commit enhances the coroutine management in the `NsmDevice` and
`SensorManagerImpl` classes by replacing raw coroutine handles with a
new `requester::Coroutine` class. Additionally, it updates the README
file to include instructions for installing missing pip modules and
apt libraries, and it adds a command to copy the latest `libmctp-externals.h`
to the common directory.

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-17810
Signed-off-by: Paweł Iwaneczko <piwaneczko@nvidia.com>