Fixes core dump issue of nsmd
'''
The issue was caused because the destroy for the co_routine was called twice.
The destroy for a co_routine may be called twice once if the task is done as soon as its created and other in destructor. Issue could be visible if the task for which co_routine was created doesn't have any co_await.

The change will assign nullptr to the destroyed co_routine, helping to avoid second destroy call in co_routine destructor.
'''

Fixes nvbug https://nvbugspro.nvidia.com/bug/5749651
signed-off-by: <ayushkumart@nvidia.com>